+++ packages/client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @alwaysmeticulous/client
 
+## 2.331.1
+
+### Patch Changes
+
+- [#12634](https://github.com/alwaysmeticulous/meticulous/pull/12634) [`c7a028b`](https://github.com/alwaysmeticulous/meticulous/commit/c7a028b28300ee2e4cef43a9a2b396828b3f76c0) Thanks [@edoardopirovano](https://github.com/edoardopirovano)! - Retry incomplete Docker image pushes and include backend container-upload diagnostics in client errors.
+
 ## 2.331.0
 
 ### Minor Changes