+++ packages/downloading-helpers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alwaysmeticulous/downloading-helpers
 
+## 2.331.1
+
+### Patch Changes
+
+- Updated dependencies [[`c7a028b`](https://github.com/alwaysmeticulous/meticulous/commit/c7a028b28300ee2e4cef43a9a2b396828b3f76c0)]:
+  - @alwaysmeticulous/client@2.331.1
+
 ## 2.331.0
 
 ### Patch Changes